# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is swagger api for Tutree.",
        "title": "Tutree Swagger API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/tutor/slots": {
            "get": {
                "description": "This function retrieves the available time slots for a tutor within a specified date range so that student can\nbook demo session with available tutor.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tutor"
                ],
                "summary": "This function retrieves the available time slots for a tutor within a specified date range.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start Date",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End Date",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}